upgrade and align with Kusama

+++ deps/ledger-zxlib/src/buffering.c
@@ -50,7 +50,7 @@ int buffering_append(uint8_t *data, int length) {
     if (ram.in_use) {
         if (ram.size - ram.pos >= length) {
             // RAM in use, append to ram if there is enough space
-            MEMCPY(ram.data + ram.pos, data, length);
+            MEMCPY(ram.data + ram.pos, data, (size_t) length);
             ram.pos += length;
         } else {
             // If RAM is not big enough copy memory to flash
@@ -66,7 +66,7 @@ int buffering_append(uint8_t *data, int length) {
     } else {
         // Flash in use, append to flash
         if (flash.size - flash.pos >= length) {
-            MEMCPY_NV(flash.data + flash.pos, data, length);
+            MEMCPY_NV(flash.data + flash.pos, data, (size_t) length);
             flash.pos += length;
         } else {
             return 0;
